resolve dependency issues
