+++ packages/superset-ui-plugin-chart-word-cloud/package.json
@@ -13,7 +13,9 @@
     "type": "git",
     "url": "git+https://github.com/apache-superset/superset-ui.git"
   },
-  "keywords": ["superset"],
+  "keywords": [
+    "superset"
+  ],
   "author": "Superset",
   "license": "Apache-2.0",
   "bugs": {
@@ -22,5 +24,16 @@
   "homepage": "https://github.com/apache-superset/superset-ui#readme",
   "publishConfig": {
     "access": "public"
+  },
+  "dependencies": {
+    "@superset-ui/chart": "^0.7.0",
+    "@superset-ui/color": "^0.7.0",
+    "@superset-ui/core": "^0.7.0",
+    "@superset-ui/translation": "^0.7.0",
+    "d3-array": "^2.0.2",
+    "d3-cloud": "^1.2.5",
+    "d3-scale": "^2.1.2",
+    "d3-selection": "^1.3.2",
+    "prop-types": "^15.6.2"
   }
 }
