+++ packages/superset-ui-plugin-chart-word-cloud/src/index.js
@@ -1,2 +1,23 @@
-const x = 1;
-export default x;
+import { t } from '@superset-ui/translation';
+import { ChartMetadata, ChartPlugin } from '@superset-ui/chart';
+import buildQuery from './buildQuery';
+import transformProps from './transformProps';
+import thumbnail from './images/thumbnail.png';
+
+const metadata = new ChartMetadata({
+  credits: ['https://github.com/jasondavies/d3-cloud'],
+  description: '',
+  name: t('Word Cloud'),
+  thumbnail,
+});
+
+export default class WordCloudChartPlugin extends ChartPlugin {
+  constructor() {
+    super({
+      buildQuery,
+      loadChart: () => import('./ReactWordCloud.js'),
+      metadata,
+      transformProps,
+    });
+  }
+}
