+++ packages/superset-ui-plugin-chart-word-cloud/test/buildQuery.test.ts
@@ -1,4 +1,4 @@
-import buildQuery from 'src/visualizations/wordcloud/buildQuery';
+import buildQuery from '../src/buildQuery';
 
 describe('WordCloud buildQuery', () => {
   const formData = {
